move navbar to page and use same instance of books
